pass calculation done and factored out
